% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisArrows.R
\name{axisArrows}
\alias{axisArrows}
\title{Axis arrows}
\usage{
axisArrows(side = 1:2, backward = FALSE, offset = 0.1, ...)
}
\arguments{
\item{side}{numeric indicating which sides(s) to put arrows on (following convention of \code{axis})}

\item{backward}{logical, should the arrow be backwards (i.e. pointing down if side is 2 or 4, or 
pointing left if side is 1 or 3)}

\item{offset}{numeric given the proportion by which to offset the start and end of the arrow from the plot edges}

\item{...}{arguments passed to function \code{arrows}}
}
\description{
Add arrows to axes to indicate direction
}
\details{
This function puts arrows in the margin space and will try to fit the arrows nicely, 
but assumes some space has been left for them (e.g. through \code{mar} argument to \code{par})
}
\examples{
plot(1)
axisArrows(side=1:2, length=0.1)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
axis, arrows, logAxis
}

