% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeCol.R
\name{cladeCol}
\alias{cladeCol}
\title{Create colors for plotting clades}
\usage{
cladeCol(phy, clades, cols, bg.col = par("fg"))
}
\arguments{
\item{phy}{the phylogeny for which to compute colors}

\item{clades}{a list of clades}

\item{cols}{a vector of colors, one for each clade}

\item{bg.col}{color for ``background'' edges (i.e. those not specified by \code{clades})}
}
\value{
A character vector of the desired colors
}
\description{
Create a vector of colors for use in \code{plot.phylo} so that designated clades have a unique color
}
\details{
Clades are specified by the names of their members, see example
}
\examples{
phy <- read.tree(text="(A,(C,G));")
plot(phy, edge.color=clade.col(phy, list("A",c("C","G")),c("red","blue")))

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
phylo, plot.phylo
}

