% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denFill.R
\name{denFill}
\alias{denFill}
\title{Density plot with confidence interval filled in}
\usage{
denFill(dist, alpha = 0.05, alternative = c("two.sided", "less", "greater"),
  col = "black", lwd = 2, transp = 0.3, add = FALSE, ...)
}
\arguments{
\item{dist}{numeric vector for which density is to be plotted}

\item{alpha}{critical value}

\item{alternative}{specify whether to plot a two-sided CI, upper or lower}

\item{col}{color for plotting}

\item{lwd}{line width}

\item{transp}{transparancy of shadded region}

\item{add}{logical, whether to add to an existing plot}

\item{...}{further arguments passed to \code{plot}}
}
\description{
Plot the interpolated density of data (most likely results from a Monte Carlo simulation 
or estimation) along with a confidence interval (specified by \code{alpha}) shadded in.
}
\details{
This function can plot desnity and quantiles for any vector and would be well suited for bootstrapping, 
permutational null models and also credible intervals from Bayesian MCMC samples
}
\examples{
x <- rnorm(100)
denFill(x)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
density, quantile
}

