% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoAxis.R
\name{paleoAxis}
\alias{paleoAxis}
\title{Plotting stratographic interval names}
\usage{
paleoAxis(side = 1, showQ = FALSE, lspace = 1.5)
}
\arguments{
\item{side}{the side on which to add an axis (following \code{axis})}

\item{showQ}{logical, should the Quaturnary be shown}

\item{lspace}{line spaces}
}
\description{
Add stratographic interval names to the axis of an existing plot
}
\details{
Stratographic intervals are taken from XXX
}
\examples{
par(mar=c(4, 4, 4, 4), mgp=c(3, 1, 0), xpd=NA)
plot(1, xlim=c(650, 0), ylim=c(650, 100), axes=FALSE)
paleoAxis(1)
paleoAxis(2)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
axis, plot
}

