% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyOrder.R
\name{prettyOrder}
\alias{prettyOrder}
\title{Pretty ordering of clades for plotting}
\usage{
prettyOrder(phy)
}
\arguments{
\item{phy}{a phylogenetic tree (class \code{phylo}) to be re-ordered}
}
\description{
Orders clades according to number of decendants for plotting purposes
}
\details{
This function function will re-order a phylogeny such that clades with the fewest
decendents will be plotted `before' (above or to the left of) clades with more decendents. 
This works recursively from the root.
}
\examples{
tre <- ape::rtree(10)
tre <- prettyOrder(tre)
ape::plot.phylo(tre)

}
\seealso{
ape::rotate
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
