% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Skewness}
\usage{
skew(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical, should NA values be removed prior to computation}
}
\value{
Numeric vector of length one giving the kurtosis value
}
\description{
Calculate the skewness of a random sample
}
\details{
Calculate the skewness of a random sample
}
\examples{
x <- rnorm(100)
skew(x)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
mean, var, kurt
}

