% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy2mat.R
\name{tidy2mat}
\alias{tidy2mat}
\title{Covert tidy data organization to matrix}
\usage{
tidy2mat(g1, g2, obs)
}
\arguments{
\item{g1}{grouping factor 1}

\item{g2}{grouping factor 2}

\item{obs}{the observed values to sum}
}
\value{
Numeric matrix with each cell the sum of observations belonging to g1 = i and g2 = j
}
\description{
Covert data organized by row (one observation per row) to a matrix where each unique group combination is summarized by the sum of its observations
}
\details{
See description
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
