% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald}
\alias{wald}
\title{Wald confidence interval}
\usage{
wald(mle, hessian, alpha = 0.05, marginal = TRUE)
}
\arguments{
\item{mle}{maximum likelihood parameter estimate}

\item{hessian}{estiamte of the Hessian at the MLE}

\item{alpha}{critical value}

\item{marginal}{logical, should marginal CI be calculated for each parameter or should an ellipse 
be calculated simultaneously for all parameters}
}
\value{
A matrix of either the upper and lower confidence limits (for \code{marginal=TRUE}) or a 
matrix with x,y coordinates fo the ellipse (for \code{marginal=FALSE})
}
\description{
Functions to calculate the Wald confidence interval of an MLE
}
\details{
Can calcuate either a Wald CI for a single parameter or multiple parameters. The \code{mle} and 
\code{hessian} arguments can be obtained from numerical optimization of the likelihood function using, 
e.g., \code{optim} or \code{nlm}
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
optim, nlm
}

