% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colAlpha.R
\name{colAlpha}
\alias{colAlpha}
\title{Add an alpha chanel to any color}
\usage{
colAlpha(col, alpha = 1)
}
\arguments{
\item{col}{the color to be modified}

\item{alpha}{the alpha level to be added}
}
\value{
A character vector of the desired colors
}
\description{
Make any color have transparancy by adding an alpha chanel
}
\details{
This function converts a color to red, blue, green components and then adds the desired alpha level
}
\examples{
## demonstrates use of the alpha argument: we can specify simple colors and use them in plots
## along with semi-transparent versions of them
cols <- c('red', 'blue', 'green')
plot(1:3, col=cols, bg=colAlpha(cols, alpha=c(0.7, 0.4, 0.1)), pch=21, cex=2, lwd=2)

}
\seealso{
hsv, rgb, col2rgb
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
