% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAxis.R
\name{logAxis}
\alias{logAxis}
\title{Log axes}
\usage{
logAxis(side, ...)
}
\arguments{
\item{side}{the side on which to add an axis (following \code{axis})}

\item{...}{arguments passed to \code{axis}}
}
\description{
Add good looking log scale axes to a plot
}
\details{
Powers of 10 receive a long tick mark and an axis label while integers that are not 
powers of 10 receive a short tick and no label
}
\examples{
plot(1, xlim=c(1, 100), xaxt='n')
logAxis(1)

}
\author{
Andy Rominger <ajrominger@gmail.com>
}
\seealso{
axis, plot
}

